Canonicalize output lemmas from ext_theory.cpp
Sometimes ext_theory.cpp was producing lemmas of the form

(A1 & A2 & A3 & A4) -> C

However, when the proof for a lemma is constructed, that lemma is looked
up as a clause, so lemmas must be clauses (or things that the theory
engine can easily turn into clauses, since the theory engine has a
limited-power clausifier).

Now the concerning lemmas from ext_theory.cpp have the form

~A1 | ~A2 | ~A3 | ~A4 | C

which is already a clause.

This change causes us to timeout (20min) on the test

regress/regress1/arith/real2int-test.smt2

in debug mode (no problem in production). Without the the change, the
benchmark is completed in 3.527792095s. Not sure why, but my hypothesis
is that the flat clause structure causes poorer splitting to happen.

I also took the liberty of adding 1-layer flattening of (->) to the
theory engine's lemma saving. That would be insufficient here because
our concerning lemmas have a conjunction for the hypothesis, but it is a
good idea in general.