bugfix: unsound SAT in --ff-solver int
Thanks for the bug report, Sorawee!

The issue was that the "unsound" mode of the range solver is suppose to
be unsound for UNSAT, but sound ford SAT. However, it was accidentally
unsound for SAT because we were treating field disequalities as integer
disequalities.

We probably didn't notice until now because != is rare in our systems.